test: split assets retry cases
